requirements.txt
streamlit
faiss-cpu
sentence-transformers
PyPDF2
python-dotenv
groq
numpy
